+++ onur/misc/log.py
@@ -1,0 +1,45 @@
+# Onur is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# Onur is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Onur. If not, see <https://www.gnu.org/licenses/>.
+
+"""Log miscellaneous information."""
+
+import logging
+
+from . import globals
+
+
+def loggin():
+    """..."""
+    logging.basicConfig(
+        filename=globals.configDir.joinpath("onur.log"),
+        encoding="utf-8",
+        level=logging.INFO,
+    )
+
+
+def info(message: str):
+    """..."""
+    loggin()
+    logging.info(message)
+
+
+def warning(message: str):
+    """..."""
+    loggin()
+    logging.warning(message)
+
+
+def error(message: str):
+    """..."""
+    loggin()
+    logging.error(message)